packages/controlbridge-core/pyproject.toml
[project]
name = "controlbridge-core"
version = "0.1.0"
description = "Core data models, catalog loading, and gap analysis engine for ControlBridge GRC tool"
readme = "README.md"
license = {text = "Apache-2.0"}
requires-python = ">=3.12"
dependencies = [
    "pydantic>=2.9",
    "pydantic-settings>=2.6",
    "pyyaml>=6.0",
    "python-dateutil>=2.9",
    "thefuzz[speedup]>=0.22",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/controlbridge_core"]
